requirements.txt
aiosqlite==0.18.0
alembic==1.10.2
anyio==3.6.2
apprise==1.3.0
asgi-lifespan==2.0.0
asyncpg==0.27.0
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
coolname==2.2.0
croniter==1.3.8
cryptography==39.0.2
dateparser==1.1.7
docker==6.0.1
fastapi==0.95.0
fsspec==2023.1.0
google-auth==2.16.2
greenlet==2.0.2
griffe==0.25.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
kubernetes==26.1.0
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.8.7
packaging==23.0
pathspec==0.11.1
pendulum==2.1.2
prefect==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.6
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==6.0
readchar==4.0.5
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
text-unidecode==1.3
toml==0.10.2
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.21.1
websocket-client==1.5.1
websockets==10.4
